api/requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
pydantic==1.9.1
passlib==1.7.4
SQLAlchemy==1.4.37
alembic==1.8.0
cryptography==37.0.2
python-jose==3.3.0
psycopg2-binary==2.9.3
pytest==7.1.2
requests==2.27.1
alembic==1.8.0
anyio==3.6.1
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
ecdsa==0.17.0
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-jose==3.3.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
